requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.4.0
beautifulsoup4==4.14.2
black==25.11.0
blinker==1.9.0
bs4==0.0.2
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
easyocr==1.7.2
executing==2.2.1
filelock==3.20.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.43.0
google-genai==1.50.1
googlemaps==4.10.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
lazy_loader==0.4
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mpmath==1.3.0
mypy_extensions==1.1.0
narwhals==2.11.0
nest-asyncio==1.6.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
opencv-python-headless==4.12.0.88
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
protobuf==6.33.1
psutil==7.1.3
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PySocks==1.7.1
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
rsa==4.9.1
scikit-image==0.25.2
scipy==1.16.3
selenium==4.38.0
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.44
stack-data==0.6.3
streamlit==1.51.0
sympy==1.14.0
tenacity==9.1.2
tifffile==2025.10.16
toml==0.10.2
torch==2.9.1
torchvision==0.24.1
tornado==6.5.2
traitlets==5.14.3
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.14
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==15.0.1
wsproto==1.3.1
anyio==4.12.0
beautifulsoup4==4.14.3
branca==0.8.2
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
folium==0.20.0
fonttools==4.61.0
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth-httplib2==0.2.1
google-genai==1.53.0
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
httplib2==0.31.0
jieba==0.42.1
kiwisolver==1.4.9
matplotlib==3.10.7
narwhals==2.12.0
numpy==2.3.5
proto-plus==1.26.1
protobuf==5.29.5
pydantic==2.12.5
pyparsing==3.2.5
RapidFuzz==3.14.3
rpds-py==0.29.0
streamlit-folium==0.25.3
tqdm==4.67.1
uritemplate==4.2.0
wordcloud==1.9.4
xyzservices==2025.11.0
